@@ -1,188 +1,168 @@
 package tk.themcbros.interiormod.blocks;
 
-import java.util.List;
-import java.util.Locale;
-import java.util.Random;
-
-import javax.annotation.Nullable;
-
 import com.google.common.collect.Lists;
-
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.IWaterLoggable;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.fluid.Fluids;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.item.ItemStack;
-import net.minecraft.state.BooleanProperty;
-import net.minecraft.state.EnumProperty;
-import net.minecraft.state.StateContainer.Builder;
-import net.minecraft.state.properties.BlockStateProperties;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.IStringSerializable;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.util.math.shapes.VoxelShapes;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorld;
-import net.minecraft.world.IWorldReader;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
-import tk.themcbros.interiormod.init.InteriorTileEntities;
+import net.minecraft.core.BlockPos;
+import net.minecraft.util.StringRepresentable;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.Shapes;
+import net.minecraft.world.phys.shapes.VoxelShape;
+import tk.themcbros.interiormod.init.InteriorBlockEntities;
 import tk.themcbros.interiormod.util.ShapeUtils;
 
-public class LampOnAStickBlock extends Block implements IWaterLoggable {
+import javax.annotation.Nullable;
+import java.util.List;
+import java.util.Locale;
 
-	public static final EnumProperty<Part> PART = EnumProperty.create("part", Part.class);
-	public static final BooleanProperty LIT = BlockStateProperties.LIT;
-	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
-	
-	private final VoxelShape SHAPE;
+public class LampOnAStickBlock extends BaseEntityBlock implements SimpleWaterloggedBlock {
 
-	public LampOnAStickBlock(Properties properties) {
-		super(properties);
-		this.setDefaultState(this.stateContainer.getBaseState().with(PART, Part.BOTTOM).with(LIT, Boolean.FALSE).with(WATERLOGGED, Boolean.FALSE));
-		this.SHAPE = this.generateShape();
-	}
-	
-	private VoxelShape generateShape() {
-		List<VoxelShape> shapes = Lists.newArrayList();
-		shapes.add(Block.makeCuboidShape(6, 0, 6, 10, 32, 10));
-		shapes.add(Block.makeCuboidShape(0, 32, 0, 16, 48, 16));
-		return ShapeUtils.combineAll(shapes);
-	}
-	
-	@Override
-	public FluidState getFluidState(BlockState state) {
-		return state.get(WATERLOGGED) ? Fluids.WATER.getDefaultState() : Fluids.EMPTY.getDefaultState();
-	}
+    public static final EnumProperty<Part> PART = EnumProperty.create("part", Part.class);
+    public static final BooleanProperty LIT = BlockStateProperties.LIT;
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
-	@SuppressWarnings("deprecation")
-	public boolean isValidPosition(BlockState state, IWorldReader worldIn, BlockPos pos) {
-//		if (state.get(PART) != Part.TOP) {
-//			return super.isValidPosition(state, worldIn, pos);
-//		} else {
-//			BlockState blockstate = worldIn.getBlockState(pos.down());
-//			if (state.getBlock() != this)
-//				return super.isValidPosition(state, worldIn, pos);
-//			return blockstate.getBlock() == this && blockstate.get(PART) == Part.BOTTOM;
-//		}
+    private final VoxelShape SHAPE;
 
-		Part part = state.get(PART);
-		if (part == Part.MIDDLE) {
-			BlockState bottomState = worldIn.getBlockState(pos.down());
-			return bottomState.getBlock() == this && bottomState.get(PART) == Part.BOTTOM;
-		} else if (part == Part.TOP) {
-			BlockState middleState = worldIn.getBlockState(pos.down());
-			return middleState.getBlock() == this && middleState.get(PART) == Part.MIDDLE;
-		} else {
-			return super.isValidPosition(state, worldIn, pos);
-		}
-	}
+    public LampOnAStickBlock(Properties properties) {
+        super(properties);
+        this.registerDefaultState(this.stateDefinition.any()
+                .setValue(PART, Part.BOTTOM).setValue(LIT, Boolean.FALSE).setValue(WATERLOGGED, Boolean.FALSE));
+        this.SHAPE = this.generateShape();
+    }
 
-	public void harvestBlock(World worldIn, PlayerEntity player, BlockPos pos, BlockState state,
-			@Nullable TileEntity te, ItemStack stack) {
-		super.harvestBlock(worldIn, player, pos, Blocks.AIR.getDefaultState(), te, stack);
-	}
+    private VoxelShape generateShape() {
+        List<VoxelShape> shapes = Lists.newArrayList();
+        shapes.add(Block.box(6, 0, 6, 10, 32, 10));
+        shapes.add(Block.box(0, 32, 0, 16, 48, 16));
+        return ShapeUtils.combineAll(shapes);
+    }
 
-	public void onBlockHarvested(World worldIn, BlockPos pos, BlockState state, PlayerEntity player) {
-		Part part = state.get(PART);
-		BlockPos bottomPos = part == Part.BOTTOM ? pos : part == Part.MIDDLE ? pos.down() : pos.down(2);
-		BlockPos middlePos = part == Part.MIDDLE ? pos : part == Part.BOTTOM ? pos.up() : pos.down();
-		BlockPos topPos = part == Part.TOP ? pos : part == Part.MIDDLE ? pos.up() : pos.up(2);
-		BlockState bottomState = worldIn.getBlockState(bottomPos);
-		BlockState middleState = worldIn.getBlockState(middlePos);
-		BlockState topState = worldIn.getBlockState(topPos);
-		if (bottomState.getBlock() == this && middleState.getBlock() == this && topState.getBlock() == this) {
-			worldIn.setBlockState(bottomPos, Blocks.AIR.getDefaultState(), 35);
-			worldIn.setBlockState(middlePos, Blocks.AIR.getDefaultState(), 35);
-			worldIn.setBlockState(topPos, Blocks.AIR.getDefaultState(), 35);
-			worldIn.playEvent(player, 2001, bottomPos, Block.getStateId(bottomState));
-			worldIn.playEvent(player, 2001, middlePos, Block.getStateId(middleState));
-			worldIn.playEvent(player, 2001, topPos, Block.getStateId(topState));
-			if (!worldIn.isRemote && !player.isCreative()) {
-				spawnDrops(bottomState, worldIn, bottomPos, (TileEntity) null, player, player.getHeldItemMainhand());
-				spawnDrops(middleState, worldIn, middlePos, (TileEntity) null, player, player.getHeldItemMainhand());
-				spawnDrops(topState, worldIn, topPos, (TileEntity) null, player, player.getHeldItemMainhand());
-			}
-		}
-	}
+    @Override
+    public RenderShape getRenderShape(BlockState p_49232_) {
+        return RenderShape.MODEL;
+    }
 
-	@Override
-	public void onBlockPlacedBy(World worldIn, BlockPos pos, BlockState state, LivingEntity placer, ItemStack stack) {
-		this.placeRestAt(worldIn, pos, 3);
-	}
+    @Override
+    public FluidState getFluidState(BlockState state) {
+        return state.getValue(WATERLOGGED) ? Fluids.WATER.defaultFluidState() : Fluids.EMPTY.defaultFluidState();
+    }
 
-	@Nullable
-	@Override
-	public BlockState getStateForPlacement(BlockItemUseContext context) {
-		BlockPos blockpos = context.getPos();
-		return blockpos.getY() < context.getWorld().getHeight() - 2
-				&& context.getWorld().getBlockState(blockpos.up()).isReplaceable(context)
-				&& context.getWorld().getBlockState(blockpos.up(2)).isReplaceable(context) ? this.getDefaultState()
-						: null;
-	}
+    @Override
+    @SuppressWarnings("deprecation")
+    public boolean canSurvive(BlockState state, LevelReader worldIn, BlockPos pos) {
+        Part part = state.getValue(PART);
+        if (part == Part.MIDDLE) {
+            BlockState bottomState = worldIn.getBlockState(pos.below());
+            return bottomState.getBlock() == this && bottomState.getValue(PART) == Part.BOTTOM;
+        } else if (part == Part.TOP) {
+            BlockState middleState = worldIn.getBlockState(pos.below());
+            return middleState.getBlock() == this && middleState.getValue(PART) == Part.MIDDLE;
+        } else {
+            return super.canSurvive(state, worldIn, pos);
+        }
+    }
 
-	@Override
-	public void tick(BlockState state, ServerWorld worldIn, BlockPos pos, Random rand) {
-		if (state.get(LIT) && !worldIn.isBlockPowered(pos)) {
-			worldIn.setBlockState(pos, state.cycleValue(LIT), 2);
-		}
+    @Override
+    public void playerDestroy(Level level, Player player, BlockPos pos, BlockState newState, @Nullable BlockEntity blockEntity, ItemStack stack) {
+        super.playerDestroy(level, player, pos, Blocks.AIR.defaultBlockState(), blockEntity, stack);
+    }
 
-	}
+    public void playerWillDestroy(Level worldIn, BlockPos pos, BlockState state, Player player) {
+        Part part = state.getValue(PART);
+        BlockPos bottomPos = part == Part.BOTTOM ? pos : part == Part.MIDDLE ? pos.below() : pos.below(2);
+        BlockPos middlePos = part == Part.MIDDLE ? pos : part == Part.BOTTOM ? pos.above() : pos.below();
+        BlockPos topPos = part == Part.TOP ? pos : part == Part.MIDDLE ? pos.above() : pos.above(2);
+        BlockState bottomState = worldIn.getBlockState(bottomPos);
+        BlockState middleState = worldIn.getBlockState(middlePos);
+        BlockState topState = worldIn.getBlockState(topPos);
+        if (bottomState.getBlock() == this && middleState.getBlock() == this && topState.getBlock() == this) {
+            worldIn.setBlock(bottomPos, Blocks.AIR.defaultBlockState(), 35);
+            worldIn.setBlock(middlePos, Blocks.AIR.defaultBlockState(), 35);
+            worldIn.setBlock(topPos, Blocks.AIR.defaultBlockState(), 35);
+            worldIn.levelEvent(player, 2001, bottomPos, Block.getId(bottomState));
+            worldIn.levelEvent(player, 2001, middlePos, Block.getId(middleState));
+            worldIn.levelEvent(player, 2001, topPos, Block.getId(topState));
+            if (!worldIn.isClientSide && !player.isCreative()) {
+                dropResources(bottomState, worldIn, bottomPos, null, player, player.getMainHandItem());
+                dropResources(middleState, worldIn, middlePos, null, player, player.getMainHandItem());
+                dropResources(topState, worldIn, topPos, null, player, player.getMainHandItem());
+            }
+        }
+    }
 
-	public void placeAt(IWorld worldIn, BlockPos pos, int flags) {
-		worldIn.setBlockState(pos, this.getDefaultState().with(PART, Part.BOTTOM), flags);
-		this.placeRestAt(worldIn, pos, flags);
-	}
+    @Override
+    public void setPlacedBy(Level worldIn, BlockPos pos, BlockState state, LivingEntity placer, ItemStack stack) {
+        this.placeRestAt(worldIn, pos, 3);
+    }
 
-	public void placeRestAt(IWorld worldIn, BlockPos pos, int flags) {
-		worldIn.setBlockState(pos.up(), this.getDefaultState().with(PART, Part.MIDDLE), flags);
-		worldIn.setBlockState(pos.up(2), this.getDefaultState().with(PART, Part.TOP), flags);
-	}
+    @Nullable
+    @Override
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        BlockPos blockpos = context.getClickedPos();
+        return blockpos.getY() < context.getLevel().getMaxBuildHeight() - 2
+                && context.getLevel().getBlockState(blockpos.above()).canBeReplaced(context)
+                && context.getLevel().getBlockState(blockpos.above(2)).canBeReplaced(context) ? this.defaultBlockState()
+                : null;
+    }
 
-	@Override
-	protected void fillStateContainer(Builder<Block, BlockState> builder) {
-		builder.add(PART, LIT, WATERLOGGED);
-	}
+    public void placeRestAt(Level worldIn, BlockPos pos, int flags) {
+        worldIn.setBlock(pos.above(), this.defaultBlockState().setValue(PART, Part.MIDDLE), flags);
+        worldIn.setBlock(pos.above(2), this.defaultBlockState().setValue(PART, Part.TOP), flags);
+    }
 
-	@Override
-	public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
-		Part part = state.get(PART);
-		return part == Part.BOTTOM ? SHAPE
-				: part == Part.MIDDLE ? SHAPE.withOffset(0, -1d, 0)
-				: part == Part.TOP ? SHAPE.withOffset(0, -2d, 0) : VoxelShapes.fullCube();
-	}
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(PART, LIT, WATERLOGGED);
+    }
 
-	@Override
-	public int getLightValue(BlockState state, IBlockReader world, BlockPos pos) {
-		return state.get(LIT) && state.get(PART) == Part.TOP ? 15 : 0;
-	}
+    @Override
+    public VoxelShape getShape(BlockState state, BlockGetter world, BlockPos pos, CollisionContext context) {
+        Part part = state.getValue(PART);
+        return part == Part.BOTTOM ? SHAPE
+                : part == Part.MIDDLE ? SHAPE.move(0, -1d, 0)
+                : part == Part.TOP ? SHAPE.move(0, -2d, 0) : Shapes.block();
+    }
 
-	@Override
-	public boolean hasTileEntity(BlockState state) {
-		return state.get(PART) == Part.TOP;
-	}
+    @Override
+    public int getLightEmission(BlockState state, BlockGetter world, BlockPos pos) {
+        return state.getValue(LIT) && state.getValue(PART) == Part.TOP ? 15 : 0;
+    }
 
-	@Override
-	public TileEntity createTileEntity(BlockState state, IBlockReader world) {
-		return state.get(PART) == Part.TOP ? InteriorTileEntities.LAMP.create() : null;
-	}
+    @Nullable
+    @Override
+    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
+        return !level.isClientSide && level.dimensionType().hasSkyLight() ? createTickerHelper(blockEntityType, InteriorBlockEntities.LAMP, NightlightLampBlock::tickEntity) : null;
+    }
 
-	public enum Part implements IStringSerializable {
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos blockPos, BlockState blockState) {
+        return InteriorBlockEntities.LAMP.create(blockPos, blockState);
+    }
 
-		BOTTOM, MIDDLE, TOP;
+    public enum Part implements StringRepresentable {
+        BOTTOM, MIDDLE, TOP;
 
-		@Override
-		public String getString() {
-			return name().toLowerCase(Locale.ROOT);
-		}
-
-	}
+        @Override
+        public String getSerializedName() {
+            return name().toLowerCase(Locale.ROOT);
+        }
+    }
 
 }
