@@ -3,29 +3,28 @@ package tk.themcbros.interiormod.blocks;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.IWaterLoggable;
-import net.minecraft.fluid.Fluids;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.state.BooleanProperty;
-import net.minecraft.state.Property;
-import net.minecraft.state.StateContainer.Builder;
-import net.minecraft.state.properties.BlockStateProperties;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorld;
-import net.minecraft.world.World;
-import tk.themcbros.interiormod.InteriorMod;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.SimpleWaterloggedBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 import tk.themcbros.interiormod.api.furniture.FurnitureType;
-import tk.themcbros.interiormod.init.InteriorTileEntities;
+import tk.themcbros.interiormod.init.InteriorBlockEntities;
 import tk.themcbros.interiormod.util.ShapeUtils;
 
+import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -33,7 +32,7 @@ import java.util.Map;
 /**
  * @author TheMCBrothers
  */
-public class TableBlock extends FurnitureBlock implements IWaterLoggable {
+public class TableBlock extends FurnitureBlock implements SimpleWaterloggedBlock {
 
     public static final BooleanProperty NORTH = BlockStateProperties.NORTH;
     public static final BooleanProperty EAST = BlockStateProperties.EAST;
@@ -52,28 +51,28 @@ public class TableBlock extends FurnitureBlock implements IWaterLoggable {
 
     public TableBlock(Properties properties) {
         super(FurnitureType.TABLE, properties);
-        this.setDefaultState(this.stateContainer.getBaseState()
-                .with(NORTH, Boolean.FALSE)
-                .with(EAST, Boolean.FALSE)
-                .with(SOUTH, Boolean.FALSE)
-                .with(WEST, Boolean.FALSE)
-                .with(WATERLOGGED, Boolean.FALSE));
-        SHAPES = this.generateShapes(this.stateContainer.getValidStates());
+        this.registerDefaultState(this.stateDefinition.any()
+                .setValue(NORTH, Boolean.FALSE)
+                .setValue(EAST, Boolean.FALSE)
+                .setValue(SOUTH, Boolean.FALSE)
+                .setValue(WEST, Boolean.FALSE)
+                .setValue(WATERLOGGED, Boolean.FALSE));
+        SHAPES = this.generateShapes(this.stateDefinition.getPossibleStates());
     }
 
     private ImmutableMap<BlockState, VoxelShape> generateShapes(ImmutableList<BlockState> states) {
-        final VoxelShape legNorthWest = (Block.makeCuboidShape(1, 0, 1, 3, 14, 3)); // LEG1
-        final VoxelShape legNorthEast = (Block.makeCuboidShape(13, 0, 1, 15, 14, 3)); // LEG2
-        final VoxelShape legSouthEast = (Block.makeCuboidShape(13, 0, 13, 15, 14, 15)); // LEG3
-        final VoxelShape legSouthWest = (Block.makeCuboidShape(1, 0, 13, 3, 14, 15)); // LEG4
-        final VoxelShape cube1 = Block.makeCuboidShape(0, 14, 0, 16, 16, 16); // PLATE
+        final VoxelShape legNorthWest = (Block.box(1, 0, 1, 3, 14, 3)); // LEG1
+        final VoxelShape legNorthEast = (Block.box(13, 0, 1, 15, 14, 3)); // LEG2
+        final VoxelShape legSouthEast = (Block.box(13, 0, 13, 15, 14, 15)); // LEG3
+        final VoxelShape legSouthWest = (Block.box(1, 0, 13, 3, 14, 15)); // LEG4
+        final VoxelShape cube1 = Block.box(0, 14, 0, 16, 16, 16); // PLATE
 
         ImmutableMap.Builder<BlockState, VoxelShape> builder = new ImmutableMap.Builder<>();
         for (BlockState state : states) {
-            boolean north = state.get(NORTH);
-            boolean east = state.get(EAST);
-            boolean south = state.get(SOUTH);
-            boolean west = state.get(WEST);
+            boolean north = state.getValue(NORTH);
+            boolean east = state.getValue(EAST);
+            boolean south = state.getValue(SOUTH);
+            boolean west = state.getValue(WEST);
 
             List<VoxelShape> shapes = Lists.newArrayList(cube1);
             if (!north && !west) {
@@ -94,30 +93,30 @@ public class TableBlock extends FurnitureBlock implements IWaterLoggable {
     }
 
     @Override
-    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+    public VoxelShape getShape(BlockState state, BlockGetter world, BlockPos pos, CollisionContext context) {
         return SHAPES.get(state);
     }
 
     @Override
-    protected void fillStateContainer(Builder<Block, BlockState> builder) {
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         builder.add(NORTH, EAST, SOUTH, WEST, WATERLOGGED);
     }
 
     @Override
-    public BlockState updatePostPlacement(BlockState stateIn, Direction facing, BlockState facingState, IWorld worldIn, BlockPos currentPos, BlockPos facingPos) {
+    public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, LevelAccessor worldIn, BlockPos currentPos, BlockPos facingPos) {
         if (!facing.getAxis().isHorizontal())
             return stateIn;
         BooleanProperty property = FACING_TO_PROPERTY.get(facing.getOpposite());
-        boolean flag = facingState.hasProperty(property) && facingState.get(property);
-        return stateIn.with(FACING_TO_PROPERTY.get(facing), flag);
+        boolean flag = facingState.hasProperty(property) && facingState.getValue(property);
+        return stateIn.setValue(FACING_TO_PROPERTY.get(facing), flag);
     }
 
     @Override
-    public BlockState getStateForPlacement(BlockItemUseContext context) {
-        World world = context.getWorld();
-        BlockPos blockPos = context.getPos();
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        Level world = context.getLevel();
+        BlockPos blockPos = context.getClickedPos();
         FluidState fluidState = world.getFluidState(blockPos);
-        boolean flag = context.hasSecondaryUseForPlayer();
+        boolean flag = context.isSecondaryUseActive();
         boolean north, east, south, west;
         if (flag) {
             north = false;
@@ -130,22 +129,24 @@ public class TableBlock extends FurnitureBlock implements IWaterLoggable {
             south = this.isTableBlock(world, blockPos.south());
             west = this.isTableBlock(world, blockPos.west());
         }
-        return this.getDefaultState().with(NORTH, north).with(EAST, east).with(SOUTH, south).with(WEST, west).with(WATERLOGGED, fluidState.getFluid() == Fluids.WATER);
+        return this.defaultBlockState().setValue(NORTH, north).setValue(EAST, east).setValue(SOUTH, south)
+                .setValue(WEST, west).setValue(WATERLOGGED, fluidState.getType() == Fluids.WATER);
     }
 
-    public boolean isTableBlock(IWorld world, BlockPos pos) {
+    public boolean isTableBlock(LevelAccessor world, BlockPos pos) {
         BlockState state = world.getBlockState(pos);
-        return state.matchesBlock(this);
+        return state.is(this);
     }
 
+    @Nullable
     @Override
-    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
-        return InteriorTileEntities.TABLE.create();
+    public BlockEntity newBlockEntity(BlockPos blockPos, BlockState blockState) {
+        return InteriorBlockEntities.TABLE.create(blockPos, blockState);
     }
 
     @Override
     public FluidState getFluidState(BlockState state) {
-        return state.get(WATERLOGGED) ? Fluids.WATER.getDefaultState() : Fluids.EMPTY.getDefaultState();
+        return state.getValue(WATERLOGGED) ? Fluids.WATER.defaultFluidState() : Fluids.EMPTY.defaultFluidState();
     }
 
 }
