@@ -1,37 +1,37 @@
 package tk.themcbros.interiormod.client.renderer;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-
-import net.minecraft.client.renderer.IRenderTypeBuffer;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRenderer;
-import net.minecraft.client.renderer.entity.EntityRendererManager;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.ITextComponent;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import tk.themcbros.interiormod.entity.SeatEntity;
 
 import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 /**
  * @author TheMCBrothers
  */
+@ParametersAreNonnullByDefault
 public class SeatRenderer extends EntityRenderer<SeatEntity> {
 
-    public SeatRenderer(EntityRendererManager renderManager) {
-        super(renderManager);
+    public SeatRenderer(EntityRendererProvider.Context context) {
+        super(context);
     }
 
     @Nonnull
     @Override
-    public ResourceLocation getEntityTexture(@Nonnull SeatEntity entity) {
+    public ResourceLocation getTextureLocation(@Nonnull SeatEntity entity) {
         return new ResourceLocation("");
     }
 
     @Override
-    protected void renderName(@Nonnull SeatEntity entityIn, @Nonnull ITextComponent displayNameIn, @Nonnull MatrixStack matrixStackIn, @Nonnull IRenderTypeBuffer bufferIn, int packedLightIn) {
-        super.renderName(entityIn, displayNameIn, matrixStackIn, bufferIn, packedLightIn);
+    protected void renderNameTag(SeatEntity seatEntity, Component component, PoseStack poseStack, MultiBufferSource bufferSource, int i) {
     }
 
     @Override
-    public void render(@Nonnull SeatEntity entityIn, float entityYaw, float partialTicks, @Nonnull MatrixStack matrixStackIn, @Nonnull IRenderTypeBuffer bufferIn, int packedLightIn) {}
+    public void render(SeatEntity seatEntity, float v, float v1, PoseStack poseStack, MultiBufferSource bufferSource, int i) {
+    }
 }
