@@ -1,13 +1,13 @@
 package tk.themcbros.interiormod.compat.jei;
 
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.inventory.IInventory;
-import net.minecraft.item.crafting.ICraftingRecipe;
-import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.item.crafting.IRecipeType;
-import net.minecraft.item.crafting.RecipeManager;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.Container;
+import net.minecraft.world.item.crafting.CraftingRecipe;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeManager;
+import net.minecraft.world.item.crafting.RecipeType;
 import tk.themcbros.interiormod.init.InteriorRecipeTypes;
 
 import java.util.*;
@@ -19,18 +19,18 @@ public class InteriorRecipes {
     private final RecipeManager recipeManager;
 
     public InteriorRecipes() {
-        ClientWorld world = Minecraft.getInstance().world;
+        ClientLevel world = Minecraft.getInstance().level;
         assert world != null;
         this.recipeManager = world.getRecipeManager();
     }
 
-    public List<ICraftingRecipe> getFurnitureCraftingRecipes() {
+    public List<CraftingRecipe> getFurnitureCraftingRecipes() {
         return new ArrayList<>(getRecipes(recipeManager, InteriorRecipeTypes.FURNITURE_CRAFTING));
     }
 
     @SuppressWarnings("unchecked")
-    private <C extends IInventory, T extends IRecipe<C>> Collection<T> getRecipes(RecipeManager recipeManager, IRecipeType<T> recipeType) {
-        Map<ResourceLocation, IRecipe<?>> recipes = recipeManager.recipes.getOrDefault(recipeType, Collections.emptyMap());
+    private <C extends Container, T extends Recipe<C>> Collection<T> getRecipes(RecipeManager recipeManager, RecipeType<T> recipeType) {
+        Map<ResourceLocation, Recipe<?>> recipes = recipeManager.recipes.getOrDefault(recipeType, Collections.emptyMap());
         return (Collection<T>) recipes.values();
     }
 }
