@@ -1,25 +1,13 @@
 package tk.themcbros.interiormod.compat.jei;
 
-import mezz.jei.api.IModPlugin;
-import mezz.jei.api.JeiPlugin;
-import mezz.jei.api.MethodsReturnNonnullByDefault;
-import mezz.jei.api.helpers.IGuiHelper;
-import mezz.jei.api.helpers.IJeiHelpers;
-import mezz.jei.api.registration.*;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import tk.themcbros.interiormod.InteriorMod;
-import tk.themcbros.interiormod.client.screen.FurnitureWorkbenchScreen;
-import tk.themcbros.interiormod.container.FurnitureWorkbenchContainer;
-import tk.themcbros.interiormod.init.InteriorBlocks;
-import tk.themcbros.interiormod.init.InteriorItems;
+import net.minecraft.MethodsReturnNonnullByDefault;
 
 /**
  * @author TheMCBrothers
  */
-@JeiPlugin
+//@JeiPlugin
 @MethodsReturnNonnullByDefault
-public class JEIPlugin implements IModPlugin {
+public class JEIPlugin/* implements IModPlugin {
 
     private final ResourceLocation PLUGIN_UID = new ResourceLocation(InteriorMod.MOD_ID, "jeiplugin");
 
@@ -55,11 +43,13 @@ public class JEIPlugin implements IModPlugin {
 
     @Override
     public void registerRecipeTransferHandlers(IRecipeTransferRegistration registration) {
-        registration.addRecipeTransferHandler(FurnitureWorkbenchContainer.class, InteriorRecipeCategoryUid.FURNITURE_CRAFTING, 1, 9, 10, 36);
+        registration.addRecipeTransferHandler(FurnitureWorkbenchMenu.class, InteriorRecipeCategoryUid.FURNITURE_CRAFTING, 1, 9, 10, 36);
     }
 
     @Override
     public void registerRecipeCatalysts(IRecipeCatalystRegistration registration) {
         registration.addRecipeCatalyst(new ItemStack(InteriorBlocks.FURNITURE_WORKBENCH), InteriorRecipeCategoryUid.FURNITURE_CRAFTING);
     }
+}
+*/ {
 }
