@@ -1,20 +1,9 @@
 package tk.themcbros.interiormod.compat.top;
 
-import mcjty.theoneprobe.api.*;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.World;
-import tk.themcbros.interiormod.api.InteriorAPI;
-import tk.themcbros.interiormod.api.furniture.FurnitureMaterial;
-import tk.themcbros.interiormod.tileentity.FurnitureTileEntity;
-
 /**
  * @author TheMCBrothers
  */
-public class TOPProviderFurniture implements IProbeInfoProvider {
+public class TOPProviderFurniture/* implements IProbeInfoProvider {
 
     static final TOPProviderFurniture INSTANCE = new TOPProviderFurniture();
 
@@ -27,8 +16,8 @@ public class TOPProviderFurniture implements IProbeInfoProvider {
     public void addProbeInfo(ProbeMode probeMode, IProbeInfo probeInfo, PlayerEntity playerEntity, World world, BlockState blockState, IProbeHitData probeHitData) {
         BlockPos pos = probeHitData.getPos();
         TileEntity tileEntity = world.getTileEntity(pos);
-        if (tileEntity instanceof FurnitureTileEntity) {
-            FurnitureTileEntity furnitureTileEntity = (FurnitureTileEntity) tileEntity;
+        if (tileEntity instanceof FurnitureBlockEntity) {
+            FurnitureBlockEntity furnitureTileEntity = (FurnitureBlockEntity) tileEntity;
             FurnitureMaterial primaryMaterial = furnitureTileEntity.getPrimaryMaterial();
             FurnitureMaterial secondaryMaterial = furnitureTileEntity.getSecondaryMaterial();
 
@@ -37,4 +26,6 @@ public class TOPProviderFurniture implements IProbeInfoProvider {
                 probeInfo.text(secondaryMaterial.getDisplayName().mergeStyle(TextFormatting.GRAY));
         }
     }
+}
+*/ {
 }
