@@ -2,11 +2,11 @@ package tk.themcbros.interiormod.data;
 
 import com.google.common.collect.Sets;
 import net.minecraft.advancements.Advancement;
-import net.minecraft.data.AdvancementProvider;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.data.DirectoryCache;
-import net.minecraft.data.IDataProvider;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.data.DataProvider;
+import net.minecraft.data.HashCache;
+import net.minecraft.data.advancements.AdvancementProvider;
+import net.minecraft.resources.ResourceLocation;
 import tk.themcbros.interiormod.InteriorMod;
 
 import java.io.IOException;
@@ -27,17 +27,17 @@ public class InteriorAdvancementProvider extends AdvancementProvider {
     }
 
     @Override
-    public void act(DirectoryCache cache) {
+    public void run(HashCache cache) {
         Path path = this.generator.getOutputFolder();
         Set<ResourceLocation> set = Sets.newHashSet();
-        Consumer<Advancement> consumer = (p_204017_3_) -> {
-            if (!set.add(p_204017_3_.getId())) {
-                throw new IllegalStateException("Duplicate advancement " + p_204017_3_.getId());
+        Consumer<Advancement> consumer = (advancement) -> {
+            if (!set.add(advancement.getId())) {
+                throw new IllegalStateException("Duplicate advancement " + advancement.getId());
             } else {
-                Path path1 = getPath(path, p_204017_3_);
+                Path path1 = getPath(path, advancement);
 
                 try {
-                    IDataProvider.save(GSON, cache, p_204017_3_.copy().serialize(), path1);
+                    DataProvider.save(GSON, cache, advancement.deconstruct().serializeToJson(), path1);
                 } catch (IOException ioexception) {
                     InteriorMod.LOGGER.error("Couldn't save advancement {}", path1, ioexception);
                 }
