@@ -1,19 +1,18 @@
 package tk.themcbros.interiormod.entity;
 
-import java.util.List;
-
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.IPacket;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-import net.minecraftforge.fml.network.NetworkHooks;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.AABB;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
 import tk.themcbros.interiormod.init.InteriorEntities;
 
 import javax.annotation.Nonnull;
+import java.util.List;
 
 /**
  * @author TheMCBrothers
@@ -21,48 +20,41 @@ import javax.annotation.Nonnull;
 public class SeatEntity extends Entity {
     private BlockPos source;
 
-    public SeatEntity(World world) {
+    public SeatEntity(Level world) {
         super(InteriorEntities.SEAT, world);
-        this.noClip = true;
+        this.noPhysics = true;
     }
 
-    private SeatEntity(World world, BlockPos source, double yOffset) {
+    private SeatEntity(Level world, BlockPos source, double yOffset) {
         this(world);
         this.source = source;
-        this.setPosition(source.getX() + 0.5, source.getY() + yOffset, source.getZ() + 0.5);
+        this.setPos(source.getX() + 0.5, source.getY() + yOffset, source.getZ() + 0.5);
     }
 
     @Override
-    protected void registerData() {
+    protected void defineSynchedData() {
     }
 
     @Override
     public void tick() {
         super.tick();
         if (source == null) {
-            source = this.getPosition();
+            source = this.blockPosition();
         }
-        if (!this.world.isRemote) {
-            if (this.getPassengers().isEmpty() || this.world.isAirBlock(source)) {
-                this.remove();
-                world.updateComparatorOutputLevel(getSource(), world.getBlockState(getSource()).getBlock());
+        if (!this.level.isClientSide) {
+            if (this.getPassengers().isEmpty() || this.level.getBlockState(source).isAir()) {
+                this.remove(false);
+                level.updateNeighborsAt(getSource(), level.getBlockState(getSource()).getBlock());
             }
         }
     }
 
     @Override
-    protected void readAdditional(@Nonnull CompoundNBT compound) {
-
+    protected void readAdditionalSaveData(CompoundTag p_20052_) {
     }
 
     @Override
-    protected void writeAdditional(@Nonnull CompoundNBT compound) {
-
-    }
-
-    @Override
-    public double getMountedYOffset() {
-        return 0.0;
+    protected void addAdditionalSaveData(CompoundTag p_20139_) {
     }
 
     public BlockPos getSource() {
@@ -70,26 +62,27 @@ public class SeatEntity extends Entity {
     }
 
     @Override
-    protected boolean canBeRidden(@Nonnull Entity entity) {
+    protected boolean canRide(@Nonnull Entity entity) {
         return true;
     }
 
     @Nonnull
     @Override
-    public IPacket<?> createSpawnPacket() {
+    public Packet<?> getAddEntityPacket() {
         return NetworkHooks.getEntitySpawningPacket(this);
     }
 
-    public static ActionResultType create(World world, BlockPos pos, double yOffset, PlayerEntity player) {
-        if (!world.isRemote) {
-            List<SeatEntity> seats = world.getEntitiesWithinAABB(SeatEntity.class, new AxisAlignedBB(pos.getX(),
+    public static InteractionResult create(Level world, BlockPos pos, double yOffset, Player player) {
+        if (!world.isClientSide) {
+            List<SeatEntity> seats = world.getEntitiesOfClass(SeatEntity.class, new AABB(pos.getX(),
                     pos.getY(), pos.getZ(), pos.getX() + 1.0, pos.getY() + 1.0, pos.getZ() + 1.0));
             if (seats.isEmpty()) {
                 SeatEntity seat = new SeatEntity(world, pos, yOffset);
-                world.addEntity(seat);
+                world.addFreshEntity(seat);
                 player.startRiding(seat, false);
             }
         }
-        return ActionResultType.SUCCESS;
+
+        return InteractionResult.SUCCESS;
     }
 }
