@@ -1,7 +1,7 @@
 package tk.themcbros.interiormod.init;
 
 import com.google.common.collect.Lists;
-import net.minecraft.item.crafting.IRecipeSerializer;
+import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -14,11 +14,11 @@ import java.util.List;
  * @author TheMCBrothers
  */
 public class InteriorRecipeSerializers {
-    private static final List<IRecipeSerializer<?>> SERIALIZERS = Lists.newArrayList();
+    private static final List<RecipeSerializer<?>> SERIALIZERS = Lists.newArrayList();
 
-    public static final IRecipeSerializer<FurnitureShapedRecipe> FURNITURE_SHAPED = register("furniture_shaped", new FurnitureShapedRecipe.Serializer());
+    public static final RecipeSerializer<FurnitureShapedRecipe> FURNITURE_SHAPED = register("furniture_shaped", new FurnitureShapedRecipe.Serializer());
 
-    private static <T extends IRecipeSerializer<?>> T register(String registryName, T serializer) {
+    private static <T extends RecipeSerializer<?>> T register(String registryName, T serializer) {
         serializer.setRegistryName(InteriorMod.getId(registryName));
         SERIALIZERS.add(serializer);
         return serializer;
@@ -27,7 +27,7 @@ public class InteriorRecipeSerializers {
     @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, modid = InteriorMod.MOD_ID)
     public static class Registration {
         @SubscribeEvent
-        public static void onItemRegistry(final RegistryEvent.Register<IRecipeSerializer<?>> event) {
+        public static void onItemRegistry(final RegistryEvent.Register<RecipeSerializer<?>> event) {
             SERIALIZERS.forEach(event.getRegistry()::register);
         }
     }
