@@ -2,6 +2,6 @@
 @ParametersAreNonnullByDefault
 package tk.themcbros.interiormod.init;
 
-import mcp.MethodsReturnNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
